requirements.txt
fake_useragent==0.1.11
Flask==2.0.2
itemadapter==0.2.0
pandas==1.2.1
protobuf==3.19.4
requests==2.22.0
Scrapy==2.5.0
selenium==3.141.0
webdriver_manager==3.3.0
google-api-core==1.30.0
google-api-python-client==2.39.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
google-cloud-bigquery==2.20.0
google-cloud-bigquery-storage==2.4.0
google-cloud-core==1.7.0
google-cloud-monitoring==0.31.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
pyarrow==4.0.1
scrapy-proxies==0.4
scrapy-user-agents==0.1.1